+++ appGUI/GUIElements.py
@@ -5487,6 +5487,9 @@ class FlatCAMActivityView(QtWidgets.QWidget):
         elif icon_kind == "Eclipse green":
             icon = icon_location + '/active_4_static.png'
             movie = icon_location + "/active_4.gif"
+        elif icon_kind == "Spinner green":
+            icon = icon_location + '/active_5_static.png'
+            movie = icon_location + "/active_5.gif"
         else:
             icon = icon_location + '/active_static.png'
             movie = icon_location + "/active.gif"
@@ -5700,7 +5703,7 @@ class AppSystemTray(QtWidgets.QSystemTrayIcon):
             self.menu_open.addSeparator()
 
             # Open Gerber ...
-            menu_opengerber = QtGui.QAction(QtGui.QIcon(self.app.resource_location + '/flatcam_icon24.png'),
+            menu_opengerber = QtGui.QAction(QtGui.QIcon(self.app.resource_location + '/open_gerber32.png'),
                                             '%s ...\t%s' % (_('Open Gerber'), _('Ctrl+G')), self)
             self.menu_open.addAction(menu_opengerber)