+++ appGUI/MainGUI.py
@@ -99,6 +99,8 @@ class MainGUI(QtWidgets.QMainWindow):
 
         FCLabel.patching_text_color = self.theme_safe_color
 
+        # self.setWindowFlags(QtCore.Qt.WindowType.FramelessWindowHint)
+
         # Divine icon pack by Ipapun @ finicons.com
 
         # #######################################################################
@@ -162,7 +164,7 @@ class MainGUI(QtWidgets.QMainWindow):
         )
 
         self.menufile_open = self.menufile.addMenu(
-            QtGui.QIcon(self.app.resource_location + '/folder32_bis.png'), '%s' % _('Open'))
+            QtGui.QIcon(self.app.resource_location + '/folder32.png'), '%s' % _('Open'))
         self.menufile_open.setToolTipsVisible(True)
 
         # Open Project ...
@@ -173,7 +175,7 @@ class MainGUI(QtWidgets.QMainWindow):
         self.menufile_open.addSeparator()
 
         # Open Gerber ...
-        self.menufileopengerber = QtGui.QAction(QtGui.QIcon(self.app.resource_location + '/flatcam_icon24.png'),
+        self.menufileopengerber = QtGui.QAction(QtGui.QIcon(self.app.resource_location + '/open_gerber32.png'),
                                                 '%s...\t%s' % (_('Open Gerber'), _('Ctrl+G')), self)
         self.menufile_open.addAction(self.menufileopengerber)
 
@@ -475,14 +477,14 @@ class MainGUI(QtWidgets.QMainWindow):
             QtGui.QIcon(self.app.resource_location + '/origin16.png'),
             '%s\t%s' % (_('Set Origin'), _('O')))
         self.menuedit_move2origin = self.menuedit.addAction(
-            QtGui.QIcon(self.app.resource_location + '/origin2_16.png'),
+            QtGui.QIcon(self.app.resource_location + '/move2origin32.png'),
             '%s\t%s' % (_('Move to Origin'), _('Shift+O')))
         self.menuedit_center_in_origin = self.menuedit.addAction(
-            QtGui.QIcon(self.app.resource_location + '/origin3_32.png'),
+            QtGui.QIcon(self.app.resource_location + '/custom_origin32.png'),
             '%s\t%s' % (_('Custom Origin'), ''))
 
         self.menueditjump = self.menuedit.addAction(
-            QtGui.QIcon(self.app.resource_location + '/jump_to16.png'),
+            QtGui.QIcon(self.app.resource_location + '/jump_to32.png'),
             '%s\t%s' % (_('Jump to Location'), _('J')))
         self.menueditlocate = self.menuedit.addAction(
             QtGui.QIcon(self.app.resource_location + '/locate16.png'),
@@ -764,7 +766,7 @@ class MainGUI(QtWidgets.QMainWindow):
             '%s\t%s' % (_("Simplification"), '')
         )
         self.geo_paint_menuitem = self.geo_editor_menu.addAction(
-            QtGui.QIcon(self.app.resource_location + '/paint20_1.png'),
+            QtGui.QIcon(self.app.resource_location + '/paint32.png'),
             '%s\t%s' % (_("Paint"), _('I'))
         )
         self.geo_transform_menuitem = self.geo_editor_menu.addAction(
@@ -1057,14 +1059,14 @@ class MainGUI(QtWidgets.QMainWindow):
         # ########################## File Toolbar# ###############################
         # ########################################################################
         self.file_open_gerber_btn = self.toolbarfile.addAction(
-            QtGui.QIcon(self.app.resource_location + '/flatcam_icon32.png'), _("Open Gerber"))
+            QtGui.QIcon(self.app.resource_location + '/open_gerber32.png'), _("Open Gerber"))
         self.file_open_excellon_btn = self.toolbarfile.addAction(
-            QtGui.QIcon(self.app.resource_location + '/drill32.png'), _("Open Excellon"))
+            QtGui.QIcon(self.app.resource_location + '/open_excellon32.png'), _("Open Excellon"))
         self.toolbarfile.addSeparator()
         self.file_open_btn = self.toolbarfile.addAction(
             QtGui.QIcon(self.app.resource_location + '/folder32.png'), _("Open Project"))
         self.file_save_btn = self.toolbarfile.addAction(
-            QtGui.QIcon(self.app.resource_location + '/project_save32.png'), _("Save project"))
+            QtGui.QIcon(self.app.resource_location + '/save_as.png'), _("Save project"))
 
         # ########################################################################
         # ########################## Edit Toolbar# ###############################
@@ -1078,7 +1080,7 @@ class MainGUI(QtWidgets.QMainWindow):
                                           QToolButton
                                           {
                                               color: black;
-                                              background-color: tomato;
+                                              background-color: blue;
                                           }
                                           """)
         self.editor_exit_btn.setIcon(QtGui.QIcon(self.app.resource_location + '/power16.png'))
@@ -1099,12 +1101,12 @@ class MainGUI(QtWidgets.QMainWindow):
         self.origin_btn = self.toolbaredit.addAction(
             QtGui.QIcon(self.app.resource_location + '/origin32.png'), _('Set Origin'))
         self.move2origin_btn = self.toolbaredit.addAction(
-            QtGui.QIcon(self.app.resource_location + '/origin2_32.png'), _('Move to Origin'))
+            QtGui.QIcon(self.app.resource_location + '/move2origin32.png'), _('Move to Origin'))
         self.center_in_origin_btn = self.toolbaredit.addAction(
-            QtGui.QIcon(self.app.resource_location + '/origin3_32.png'), _('Custom Origin'))
+            QtGui.QIcon(self.app.resource_location + '/custom_origin32.png'), _('Custom Origin'))
 
         self.jmp_btn = self.toolbaredit.addAction(
-            QtGui.QIcon(self.app.resource_location + '/jump_to16.png'), _('Jump to Location'))
+            QtGui.QIcon(self.app.resource_location + '/jump_to32.png'), _('Jump to Location'))
         self.locate_btn = self.toolbaredit.addAction(
             QtGui.QIcon(self.app.resource_location + '/locate32.png'), _('Locate in Object'))
 
@@ -1140,7 +1142,7 @@ class MainGUI(QtWidgets.QMainWindow):
         # ########################## Tools Toolbar# ##############################
         # ########################################################################
         self.drill_btn = self.toolbarplugins.addAction(
-            QtGui.QIcon(self.app.resource_location + '/extract_drill32.png'), _("Drilling"))
+            QtGui.QIcon(self.app.resource_location + '/drill32.png'), _("Drilling"))
         self.mill_btn = self.toolbarplugins.addAction(
             QtGui.QIcon(self.app.resource_location + '/milling_tool32.png'), _("Milling"))
         self.level_btn = self.toolbarplugins.addAction(
@@ -1151,13 +1153,13 @@ class MainGUI(QtWidgets.QMainWindow):
         self.toolbarplugins.addSeparator()
 
         self.isolation_btn = self.toolbarplugins.addAction(
-            QtGui.QIcon(self.app.resource_location + '/iso_16.png'), _("Isolation"))
+            QtGui.QIcon(self.app.resource_location + '/iso32.png'), _("Isolation"))
         self.follow_btn = self.toolbarplugins.addAction(
             QtGui.QIcon(self.app.resource_location + '/follow32.png'), _("Follow"))
         self.ncc_btn = self.toolbarplugins.addAction(
             QtGui.QIcon(self.app.resource_location + '/ncc32.png'), _("NCC"))
         self.paint_btn = self.toolbarplugins.addAction(
-            QtGui.QIcon(self.app.resource_location + '/paint20_1.png'), _("Paint"))
+            QtGui.QIcon(self.app.resource_location + '/paint32.png'), _("Paint"))
 
         self.toolbarplugins.addSeparator()
 
@@ -1188,7 +1190,7 @@ class MainGUI(QtWidgets.QMainWindow):
         self.punch_btn = self.toolbarplugins.addAction(
             QtGui.QIcon(self.app.resource_location + '/punch32.png'), _("Punch Gerber"))
         self.calculators_btn = self.toolbarplugins.addAction(
-            QtGui.QIcon(self.app.resource_location + '/calculator24.png'), _("Calculators"))
+            QtGui.QIcon(self.app.resource_location + '/calculator32.png'), _("Calculators"))
 
         self.toolbarplugins.addSeparator()
 
@@ -1267,7 +1269,7 @@ class MainGUI(QtWidgets.QMainWindow):
         self.geo_add_buffer_btn = self.geo_edit_toolbar.addAction(
             QtGui.QIcon(self.app.resource_location + '/buffer16-2.png'), _('Add Buffer'))
         self.geo_add_paint_btn = self.geo_edit_toolbar.addAction(
-            QtGui.QIcon(self.app.resource_location + '/paint20_1.png'), _('Paint Shape'))
+            QtGui.QIcon(self.app.resource_location + '/paint32.png'), _('Paint Shape'))
         self.geo_eraser_btn = self.geo_edit_toolbar.addAction(
             QtGui.QIcon(self.app.resource_location + '/eraser26.png'), _('Eraser'))
 
@@ -1355,7 +1357,7 @@ class MainGUI(QtWidgets.QMainWindow):
 
         # Snap GRID toolbar is always active to facilitate usage of measurements done on GRID
         self.grid_snap_btn = self.grid_toolbar.addAction(
-            QtGui.QIcon(self.app.resource_location + '/grid32.png'), _('Snap to grid'))
+            QtGui.QIcon(self.app.resource_location + '/grid18.png'), _('Snap to grid'))
         self.grid_gap_x_entry = FCEntry2()
         self.grid_gap_x_entry.setMaximumWidth(70)
         self.grid_gap_x_entry.setToolTip(_("Grid X snapping distance"))
@@ -1396,7 +1398,7 @@ class MainGUI(QtWidgets.QMainWindow):
         # ########################################################################
         self.axis_status_label = FCLabel()
         self.axis_status_label.setToolTip(_("Toggle the display of axis on canvas"))
-        self.axis_status_label.setPixmap(QtGui.QPixmap(self.app.resource_location + '/axis16.png'))
+        self.axis_status_label.setPixmap(QtGui.QPixmap(self.app.resource_location + '/axis18.png'))
         self.status_toolbar.addWidget(self.axis_status_label)
         self.status_toolbar.addWidget(FCLabel(" "))
 
@@ -1408,7 +1410,7 @@ class MainGUI(QtWidgets.QMainWindow):
 
         self.hud_label = FCLabel()
         self.hud_label.setToolTip(_("HUD (Heads up display)"))
-        self.hud_label.setPixmap(QtGui.QPixmap(self.app.resource_location + '/hud16.png'))
+        self.hud_label.setPixmap(QtGui.QPixmap(self.app.resource_location + '/hud18.png'))
         self.status_toolbar.addWidget(self.hud_label)
         self.status_toolbar.addWidget(FCLabel(" "))
 
@@ -1846,7 +1848,7 @@ class MainGUI(QtWidgets.QMainWindow):
         self.draw_buffer = self.g_editor_cmenu.addAction(
             QtGui.QIcon(self.app.resource_location + '/buffer16-2.png'), _("Buffer"))
         self.draw_paint = self.g_editor_cmenu.addAction(
-            QtGui.QIcon(self.app.resource_location + '/paint20_1.png'), _("Paint"))
+            QtGui.QIcon(self.app.resource_location + '/paint32.png'), _("Paint"))
         self.draw_eraser = self.g_editor_cmenu.addAction(
             QtGui.QIcon(self.app.resource_location + '/eraser26.png'), _("Eraser"))
         self.g_editor_cmenu.addSeparator()
@@ -1904,7 +1906,7 @@ class MainGUI(QtWidgets.QMainWindow):
         self.e_editor_cmenu = self.popMenu.addMenu(
             QtGui.QIcon(self.app.resource_location + '/drill32.png'), _("Exc Editor"))
         self.drill = self.e_editor_cmenu.addAction(
-            QtGui.QIcon(self.app.resource_location + '/drill32.png'), _("Add Drill"))
+            QtGui.QIcon(self.app.resource_location + '/open_excellon32.png'), _("Add Drill"))
         self.drill_array = self.e_editor_cmenu.addAction(
             QtGui.QIcon(self.app.resource_location + '/addarray32.png'), _("Add Drill Array"))
         self.e_editor_cmenu.addSeparator()
@@ -1930,7 +1932,7 @@ class MainGUI(QtWidgets.QMainWindow):
         self.popmenu_numeric_move = self.popMenu.addAction(
             QtGui.QIcon(self.app.resource_location + '/move32_bis.png'), _("Num Move"))
         self.popmenu_move2origin = self.popMenu.addAction(
-            QtGui.QIcon(self.app.resource_location + '/origin2_32.png'), _("Move2Origin"))
+            QtGui.QIcon(self.app.resource_location + '/move2origin32.png'), _("Move2Origin"))
         self.popmenu_move = self.popMenu.addAction(
             QtGui.QIcon(self.app.resource_location + '/move32.png'), _("Move"))
         self.popmenu_properties = self.popMenu.addAction(
@@ -2453,14 +2455,14 @@ class MainGUI(QtWidgets.QMainWindow):
         # ##################### File Toolbar #####################################
         # ########################################################################
         self.file_open_gerber_btn = self.toolbarfile.addAction(
-            QtGui.QIcon(self.app.resource_location + '/flatcam_icon32.png'), _("Open Gerber"))
+            QtGui.QIcon(self.app.resource_location + '/open_gerber32.png'), _("Open Gerber"))
         self.file_open_excellon_btn = self.toolbarfile.addAction(
-            QtGui.QIcon(self.app.resource_location + '/drill32.png'), _("Open Excellon"))
+            QtGui.QIcon(self.app.resource_location + '/open_excellon32.png'), _("Open Excellon"))
         self.toolbarfile.addSeparator()
         self.file_open_btn = self.toolbarfile.addAction(
             QtGui.QIcon(self.app.resource_location + '/folder32.png'), _("Open Project"))
         self.file_save_btn = self.toolbarfile.addAction(
-            QtGui.QIcon(self.app.resource_location + '/project_save32.png'), _("Save Project"))
+            QtGui.QIcon(self.app.resource_location + '/save_as.png'), _("Save Project"))
 
         # ########################################################################
         # ######################### Edit Toolbar #################################
@@ -2474,7 +2476,7 @@ class MainGUI(QtWidgets.QMainWindow):
                                           QToolButton
                                           {
                                               color: black;
-                                              background-color: tomato;
+                                              background-color: blue;
                                           }
                                           """)
         self.editor_exit_btn.setIcon(QtGui.QIcon(self.app.resource_location + '/power16.png'))
@@ -2495,12 +2497,12 @@ class MainGUI(QtWidgets.QMainWindow):
         self.origin_btn = self.toolbaredit.addAction(
             QtGui.QIcon(self.app.resource_location + '/origin32.png'), _('Set Origin'))
         self.move2origin_btn = self.toolbaredit.addAction(
-            QtGui.QIcon(self.app.resource_location + '/origin2_32.png'), _('Move to Origin'))
+            QtGui.QIcon(self.app.resource_location + '/move2origin32.png'), _('Move to Origin'))
         self.center_in_origin_btn = self.toolbaredit.addAction(
-            QtGui.QIcon(self.app.resource_location + '/origin3_32.png'), _('Custom Origin'))
+            QtGui.QIcon(self.app.resource_location + '/custom_origin32.png'), _('Custom Origin'))
 
         self.jmp_btn = self.toolbaredit.addAction(
-            QtGui.QIcon(self.app.resource_location + '/jump_to16.png'), _('Jump to Location'))
+            QtGui.QIcon(self.app.resource_location + '/jump_to32.png'), _('Jump to Location'))
         self.locate_btn = self.toolbaredit.addAction(
             QtGui.QIcon(self.app.resource_location + '/locate32.png'), _('Locate in Object'))
 
@@ -2534,7 +2536,7 @@ class MainGUI(QtWidgets.QMainWindow):
         # ######################### Tools Toolbar #################################
         # #########################################################################
         self.drill_btn = self.toolbarplugins.addAction(
-            QtGui.QIcon(self.app.resource_location + '/extract_drill32.png'), _("Drilling"))
+            QtGui.QIcon(self.app.resource_location + '/drill32.png'), _("Drilling"))
         self.mill_btn = self.toolbarplugins.addAction(
             QtGui.QIcon(self.app.resource_location + '/milling_tool32.png'), _("Milling"))
         self.level_btn = self.toolbarplugins.addAction(
@@ -2545,13 +2547,13 @@ class MainGUI(QtWidgets.QMainWindow):
         self.toolbarplugins.addSeparator()
 
         self.isolation_btn = self.toolbarplugins.addAction(
-            QtGui.QIcon(self.app.resource_location + '/iso_16.png'), _("Isolation"))
+            QtGui.QIcon(self.app.resource_location + '/iso32.png'), _("Isolation"))
         self.follow_btn = self.toolbarplugins.addAction(
             QtGui.QIcon(self.app.resource_location + '/follow32.png'), _("Follow"))
         self.ncc_btn = self.toolbarplugins.addAction(
             QtGui.QIcon(self.app.resource_location + '/ncc32.png'), _("NCC"))
         self.paint_btn = self.toolbarplugins.addAction(
-            QtGui.QIcon(self.app.resource_location + '/paint20_1.png'), _("Paint"))
+            QtGui.QIcon(self.app.resource_location + '/paint32.png'), _("Paint"))
 
         self.toolbarplugins.addSeparator()
 
@@ -2582,7 +2584,7 @@ class MainGUI(QtWidgets.QMainWindow):
         self.punch_btn = self.toolbarplugins.addAction(
             QtGui.QIcon(self.app.resource_location + '/punch32.png'), _("Punch Gerber"))
         self.calculators_btn = self.toolbarplugins.addAction(
-            QtGui.QIcon(self.app.resource_location + '/calculator24.png'), _("Calculators"))
+            QtGui.QIcon(self.app.resource_location + '/calculator32.png'), _("Calculators"))
 
         self.toolbarplugins.addSeparator()
 
@@ -2659,7 +2661,7 @@ class MainGUI(QtWidgets.QMainWindow):
         self.geo_add_buffer_btn = self.geo_edit_toolbar.addAction(
             QtGui.QIcon(self.app.resource_location + '/buffer16-2.png'), _('Add Buffer'))
         self.geo_add_paint_btn = self.geo_edit_toolbar.addAction(
-            QtGui.QIcon(self.app.resource_location + '/paint20_1.png'), _('Paint Shape'))
+            QtGui.QIcon(self.app.resource_location + '/paint32.png'), _('Paint Shape'))
         self.geo_eraser_btn = self.geo_edit_toolbar.addAction(
             QtGui.QIcon(self.app.resource_location + '/eraser26.png'), _('Eraser'))